+++ src/rules/motion-token-use/index.js
@@ -36,7 +36,7 @@ const defaultOptions = {
   testOnlyTarget: undefined
 };
 
-export default function rule(primaryOptions, secondaryOptions) {
+export default function rule(primaryOptions, secondaryOptions, context) {
   const options = parseOptions(secondaryOptions, defaultOptions);
 
   return async (root, result) => {
@@ -66,6 +66,14 @@ export default function rule(primaryOptions, secondaryOptions) {
       return;
     }
 
-    await checkRule(root, result, ruleName, options, messages, getMotionInfo);
+    await checkRule(
+      root,
+      result,
+      ruleName,
+      options,
+      messages,
+      getMotionInfo,
+      context
+    );
   };
 }