+++ src/rules/layout-token-use/utils/fixes.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { fixUsingMap } from "../../../utils/fixUsingMap";
+import { fixUsingMap } from "../../../utils/fix-utils";
 
 export const fixes = [
   {